Arbeidet med endringer i fagsaker for ducks.
