+++ src/ducks/fagsaker/index.js
@@ -1,0 +1,15 @@
+import * as fagsakActions from './actions';
+import * as fagsakOperations from './operations';
+import * as fagsakTypes from './types';
+import * as fagsakSelectors from './selectors';
+
+import fagsakReducers from './reducers';
+
+export {
+  fagsakActions,
+  fagsakOperations,
+  fagsakTypes,
+  fagsakSelectors,
+};
+
+export default fagsakReducers;
