+++ src/ducks/fagsaker/types.js
@@ -1,0 +1,3 @@
+export const OK = 'fagsaker/OK';
+export const FEILET = 'fagsaker/FEILET';
+export const PENDING = 'fagsaker/PENDING';
